[development/graphics] some cleanup of the graphics code, mostly around the ObjectWithDestructionCallback -> ScopedPoolObject rename

+++ source/1_engine/rex_engine/src/gfx/system/command_allocator_pool.cpp
@@ -49,7 +49,7 @@ namespace rex
     }
 
     ScopedFencedAllocator::ScopedFencedAllocator()
-      : ScopedFencedAllocator(-1, nullptr, nullptr)
+      : ScopedFencedAllocator(static_cast<u64>(-1), nullptr, nullptr)
     {}
 
     ScopedFencedAllocator::ScopedFencedAllocator(u64 initialFenceValue, FencedAllocator* alloc, GrowingPool<FencedAllocator>* pool)
